+++ app/src/main/java/com/example/menumasterapp/presentation/auth/component/CustomPasswordTextField.kt
@@ -27,9 +27,9 @@ import androidx.compose.ui.text.input.PasswordVisualTransformation
 import androidx.compose.ui.text.input.VisualTransformation
 import androidx.compose.ui.unit.dp
 import com.example.menumasterapp.R
-import com.example.menumasterapp.ui.theme.Grey
-import com.example.menumasterapp.ui.theme.Primary
-import com.example.menumasterapp.ui.theme.Typography
+import com.example.menumasterapp.presentation.ui.theme.Grey
+import com.example.menumasterapp.presentation.ui.theme.Primary
+import com.example.menumasterapp.presentation.ui.theme.Typography
 
 @Composable
 fun CustomPasswordTextField(
