+++ app/src/main/java/com/example/menumasterapp/presentation/auth/component/CustomTextField.kt
@@ -19,9 +19,9 @@ import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.text.input.KeyboardType
 import androidx.compose.ui.unit.dp
 import com.example.menumasterapp.R
-import com.example.menumasterapp.ui.theme.Grey
-import com.example.menumasterapp.ui.theme.Primary
-import com.example.menumasterapp.ui.theme.Typography
+import com.example.menumasterapp.presentation.ui.theme.Grey
+import com.example.menumasterapp.presentation.ui.theme.Primary
+import com.example.menumasterapp.presentation.ui.theme.Typography
 
 @Composable
 fun CustomTextField(
