+++ app/src/main/java/com/example/menumasterapp/presentation/auth/register/RegisterViewModel.kt
@@ -3,7 +3,6 @@ package com.example.menumasterapp.presentation.auth.register
 import android.util.Patterns
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.example.menumasterapp.data.remote.repository.AuthRepositoryImpl
 import com.example.menumasterapp.domain.repository.AuthRepository
 import com.example.menumasterapp.presentation.auth.AuthState
 import dagger.hilt.android.lifecycle.HiltViewModel
