+++ app/src/main/java/com/example/menumasterapp/presentation/auth/register/component/AcceptedTermsSection.kt
@@ -8,9 +8,9 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import com.example.menumasterapp.ui.theme.LightBlack
-import com.example.menumasterapp.ui.theme.Primary
-import com.example.menumasterapp.ui.theme.Typography
+import com.example.menumasterapp.presentation.ui.theme.LightBlack
+import com.example.menumasterapp.presentation.ui.theme.Primary
+import com.example.menumasterapp.presentation.ui.theme.Typography
 
 @Composable
 fun AcceptedTermsSection(
